+++ src/main/java/com/epam/jwd/model/Person.java
@@ -6,10 +6,10 @@ public class Person extends AbstractBaseEntity {
 
     private final String login;
     private final String password;
-    private final Double balance;
+    private final Integer balance;
     private final Role role;
 
-    public Person(Long id, String login, String password, Double balance, Role role) {
+    public Person(Long id, String login, String password, Integer balance, Role role) {
         super(id);
         this.login = login;
         this.password = password;
@@ -17,7 +17,7 @@ public class Person extends AbstractBaseEntity {
         this.role = role;
     }
 
-    public Person(String login, String password, Double balance) {
+    public Person(String login, String password, Integer balance) {
         this(null, login, password, balance, Role.USER);
     }
 
@@ -29,7 +29,7 @@ public class Person extends AbstractBaseEntity {
         return password;
     }
 
-    public Double getBalance() {
+    public Integer getBalance() {
         return balance;
     }
